Use a producer to stream back responses
The problem with dumping all of the json response into the Request object at
once is that doing so starts the timeout for the next request to be received:
so if it takes longer than 60s to stream back the response to the client, the
client never gets it.

The correct solution is to use Producer; then the timeout is only started once
all of the content is sent over the TCP connection.

However, we then need to hook the completion logging in via notifyFinish (which
is a much better way of doing things anyway), in order to correctly read the
response length.