Refactor to allow event signing to depend on room version
In the futures we'll want to be able to change the event formats, and so
we will want to be able to change the way we sign and hash events.

In particular we want to be able to change event IDs to be the same as
their reference hashes, and so the reference hashes have to remove the
event ID field before hashing.