Refactor StateFilter to be cleaner
This changes the internal representation of `StateFilter` to remove the
concept of `filtered_types` in favour of a `types` dict and a flag to
indicate whether or not to fetch types not in the `types` dict.

Ideally in the future we should be able to rewrite our caching so that
we cache the filter used when fetching state from the DB. Then when we
go fetch more state we can compare the given filter with the filter of
the cached result to see if the given is a subset of the cached filter.
This would avoid the convolutions of maintaining separate caches.