Downgrade to debhelper compat 9
We're supposed to match the debhelper compat level with our debhelper
dependency. Xenial only has debhelper 9, so we should use that.

I think the reason for this is basically that compat level 10 was considered
unstable until debhelper 10 was released - basically we're getting some
experimental pre-alpha version of compat level 10 by asking for it in
xenial. It may be fine, or it may not. Let's play safe.