INSTALL.md: Add repo-key.asc fingerprint verification
For the apt-secure(8) infrastructure, it is critical keys added manually
via apt-key(8) must be verified. In an attempt to not undermine that
security (in an unlikely event the matrix.org server would be
compromised), print the fingerprint of the repository key while
importing it to apt(8)'s keyrings.

It is also implied in apt-key(8) man page that the "add" command should
not be used, but instead keyrings added to the /etc/apt/trusted.gpg.d/
directory.

Add a copy of the fingerprint in the instructions.

Silence curl's output while here, for ease of verification.

Requires GnuPG version 2.1.14 or later. No transitional package (gnupg2)
is required in Debian, since Debian 9 (stretch).