Collect outgoing read-receipts into buckets
In order to reduce the number of outgoing federation transactions, we want to
aggregate read-receipts, since there is little harm if outgoing RRs are delayed
by a few seconds, and we can dramatically reduce the number of transactions by
doing so.

This change introduces the concept of EDU 'buckets'; currently we have the
'Instant' bucket and the 'Delayed' bucket.

Fixes #4730, #3951.