Fix sync bug when accepting invites
*Hopefully* this will fix #4422: essentially we need to make sure that the
cache on `get_rooms_for_user_with_stream_ordering` is invalidated *before* the
SyncHandler is notified for the new events. Doing so on the `events` stream
rather than the `caches` stream achieves this, and should be safe because the
only time we need to invalidate this cache is when we persist an event.